fix deploy usesearchparms
